feat: Add PrincipalLayout component and update related files
The PrincipalLayout component is added to the layouts folder. This component includes a header, outlet, and footer. The index.ts file in the layouts folder is updated to export the PrincipalLayout component. The ContactForm.tsx file in the views folder is modified to remove the Header component. The HomePage.tsx file in the views folder is modified to remove the Header and Footer components. The Footer.tsx file in the components folder is updated to include links to the about-us and contact pages. The main.tsx file is updated to import and use the PrincipalLayout component for the root path.